roundTo to Double

+++ src/RoundTo.hs
@@ -3,5 +3,5 @@ module RoundTo
     ) where
 
 import Numeric
-roundTo :: RealFloat a => a -> Int -> Float
-roundTo floatNum numOfDecimals = read (showFFloat (Just numOfDecimals) floatNum "") :: Float
+roundTo :: RealFloat a => a -> Int -> Double
+roundTo floatNum numOfDecimals = read (showFFloat (Just numOfDecimals) floatNum "") :: Double
